+++ src/TwigExtension.php
@@ -1,20 +1,19 @@
 <?php
+
 /**
- * Zend Framework (http://framework.zend.com/)
- *
- * @see       https://github.com/zendframework/zend-expressive for the canonical source repository
- * @copyright Copyright (c) 2015 Zend Technologies USA Inc. (http://www.zend.com)
- * @license   https://github.com/zendframework/zend-expressive/blob/master/LICENSE.md New BSD License
+ * @see       https://github.com/mezzio/mezzio-twigrenderer for the canonical source repository
+ * @copyright https://github.com/mezzio/mezzio-twigrenderer/blob/master/COPYRIGHT.md
+ * @license   https://github.com/mezzio/mezzio-twigrenderer/blob/master/LICENSE.md New BSD License
  */
 
-namespace Zend\Expressive\Twig;
+namespace Mezzio\Twig;
 
+use Mezzio\Router\RouterInterface;
 use Twig_Extension;
 use Twig_SimpleFunction;
-use Zend\Expressive\Router\RouterInterface;
 
 /**
- * Twig extension for rendering URLs and assets URLs from Expressive.
+ * Twig extension for rendering URLs and assets URLs from Mezzio.
  *
  * @author Geert Eltink (https://xtreamwayz.github.io)
  */
@@ -55,7 +54,7 @@ class TwigExtension extends Twig_Extension
      */
     public function getName()
     {
-        return 'zend-expressive';
+        return 'mezzio';
     }
 
     /**