+++ CHANGELOG.md
@@ -18,7 +18,7 @@ All notable changes to this project will be documented in this file, in reverse 
 
 ### Fixed
 
-- [#3](https://github.com/zendframework/zend-expressive-twigrenderer/pull/3)
+- [zendframework/zend-expressive-twigrenderer#3](https://github.com/zendframework/zend-expressive-twigrenderer/pull/3)
   updates the `renderAssetUrl()` method of the `TwigExtension` to mask
   versioning if it's empty (while also allowing zero versions).
 
@@ -38,7 +38,7 @@ All notable changes to this project will be documented in this file, in reverse 
 
 ### Fixed
 
-- Updated to zend-expressive RC1.
+- Updated to mezzio RC1.
 - Added branch-alias of dev-master to 1.0-dev.
 
 ## 0.1.0 - 2015-10-10